+++ terminice/lib/src/guides/hotkey_guide.dart
@@ -48,7 +48,6 @@ extension HotkeyGuideExtensions on Terminice {
       final widgetFrame = FrameView(
         title: title,
         theme: theme,
-        hintStyle: HintStyle.none,
       );
 
       widgetFrame.render(out, (ctx) {